+++ fractal_server/app/models/state.py
@@ -6,12 +6,13 @@ from sqlalchemy import Column
 from sqlalchemy.types import DateTime
 from sqlalchemy.types import JSON
 from sqlmodel import Field
+from sqlmodel import SQLModel
 
 from ...common.schemas import _StateBase
 from ...utils import get_timestamp
 
 
-class State(_StateBase, table=True):
+class State(_StateBase, SQLModel, table=True):
     """
     Store arbitrary data in the database