+++ CHANGELOG.md
@@ -16,6 +16,10 @@
     * Improve error message for missing package manifest (\#704).
     * Remove private/public options for task collection (\#704).
     * Improve behavior when task-collection folder already exists (\#704).
+    * Update `_TaskCollectPip` methods, attributes and properties (\#719).
+    * Set `Task.version` during task collection (\#719).
+    * Make `Task.source` task-specific rather than package-specific (\#719).
+    * When importing a workflow, only use tasks' `source` values, instead of `(source,name)` pairs (\#719).
 * Job execution:
     * Add `FractalSlurmExecutor.shutdown` and corresponding endpoint (\#631, \#691, \#696).
     * In `FractalSlurmExecutor`, make `working_dir*` attributes required (\#679).
@@ -31,7 +35,10 @@
     * Remove `fastapi-users-db-sqlmodel` dependency (\#660).
     * Make coverage measure more accurate (\#676) and improve coverage (\#678).
     * Require pydantic version to be `>=1.10.8` (\#711, \#713).
-    * Include multiple `fractal-common` updates (\#705).
+    * Include multiple `fractal-common` updates (\#705, \#719).
+    * Add test equivalent to `alembic check` (\#722).
+    * Update `poetry.lock` to address security alerts (\#723).
+    * Remove `sqlmodel` from `fractal-common`, and declare database models with multiple inheritance (\#710).
 
 # 1.2.5